Add files via upload
The investigation of sector performance and way to invest during crisis in Thailand stock market. 

crisis period identify by 
1. 1997 East Asian crisis
2. 2008 Global financial crisis
3. 2020 Covid-19 crisis

The result indicate that Food and Healthcare sector are among the best sector during the crisis. 
However, because of stock overall decline , the return is not positive, but this can be method of identifying the stock can used in portfolio allocation to these sector stocks.

The full paper is "Preventive Sectors and Risk Finance_finals_pdf" in this repository